Refactor metadata on pages
