+++ lib/ssr.ts
@@ -205,3 +205,106 @@ export const ssrPrepParams = async (params: SSRParams, searchParams: SSRSearchPa
     console.log("==> common SSR", JSON.stringify({ teamName, teamid, athleteUUId, tab, view, fallback }));
     return { userInfo, dark, view, tab, rtab, fallback, fbclid, utm_content, bot, isMobile, story, findexarxid, m, league, pagetype, teamid, name, athleteUUId, teamName, ua };
 }
+
+export async function generateMetadata(
+    { searchParams }: { searchParams: SSRSearchParams },
+    parent: ResolvingMetadata
+): Promise<Metadata> {
+    // Read route params
+    const { id, story, tab, view, m, s = '0' } = await searchParams as any;
+    //console.log("META searchParams", { id, story, tab, view, m, s });
+
+    let findexarxid = id || "";
+
+    let amention, astory;
+    if (findexarxid) {
+        amention = await getAMention({ type: "AMention", findexarxid });
+    }
+    if (story) {
+        astory = await getASlugStory({ type: "ASlugStory", slug: story });
+    }
+    if (m) {
+        astory = await getASlugStory({ type: "ASlugStory", m });
+    }
+
+    const { summary: amentionSummary = "", league: amentionLeague = "", type = "", team: amentionTeam = "", teamName: amentionTeamName = "", name: amentionPlayer = "", image: amentionImage = "", date: amentionDate = "" } = amention || {};
+    let {
+        title: astoryTitle = "",
+        site_name: astorySite_Name = "",
+        authors: astoryAuthors = "",
+        digest: astoryDigest = "",
+        image: astoryImage = "",
+        ogimage: astoryOgImage = "",
+        createdTime: astoryDate = "",
+        mentions: mentions = [],
+        image_width = 1200,
+        image_height = 1200
+    } = astory || {};
+
+    const astoryImageOgUrl = astoryOgImage ? astoryOgImage : astoryImage ? `${process.env.NEXT_PUBLIC_SERVER}/api/og.png/${encodeURIComponent(astoryImage)}/${encodeURIComponent(astorySite_Name)}/${image_width}/${image_height}` : ``;
+
+    // Prepare meta data for amention
+    let ogUrl = '';
+    if (amention && amentionLeague && amentionTeam && amentionPlayer) {
+        ogUrl = `${process.env.NEXT_PUBLIC_SERVER}/${amentionLeague}/team/${amentionTeam}/player/${amentionPlayer}?id=${findexarxid}`;
+    } else if (amention && amentionLeague && amentionTeam) {
+        ogUrl = `${process.env.NEXT_PUBLIC_SERVER}/${amentionLeague}/team/${amentionTeam}?id=${findexarxid}`;
+    } else if (amention && amentionLeague) {
+        ogUrl = `${process.env.NEXT_PUBLIC_SERVER}/${amentionLeague}?id=${findexarxid}`;
+    } else if (amention) {
+        ogUrl = `${process.env.NEXT_PUBLIC_SERVER}/?id=${findexarxid}`;
+    } else {
+        ogUrl = `${process.env.NEXT_PUBLIC_SERVER}`;
+    }
+
+    let ogTarget = '';
+    if (amention && amentionLeague && amentionTeam && amentionPlayer && type == 'person') {
+        ogTarget = `${amentionPlayer} of ${amentionTeamName}`;
+    } else if (amention && amentionLeague && amentionTeam) {
+        ogTarget = `${amentionTeamName} on ${process.env.NEXT_PUBLIC_APP_NAME}`;
+    }
+
+    let ogDescription = amentionSummary || "Interactive Sports Knowledge for Fantasy Sports Fans.";
+    let ogImage = astoryImageOgUrl || '/q-logo-og-1200.png';
+    if (!astoryImageOgUrl) image_height = 630;
+    let ogTitle = ogTarget || `Qwiket AI`;
+    if (astory) {
+        ogUrl = story ? `${process.env.NEXT_PUBLIC_SERVER}/?story=${story}` : `${process.env.NEXT_PUBLIC_SERVER}/`;
+        ogTitle = astoryTitle;
+        ogDescription = astoryDigest.replaceAll('<p>', '').replaceAll('</p>', "\n\n");
+        ogImage = astoryImageOgUrl;
+    }
+    const noindex = 1;
+
+    return {
+        title: ogTitle,
+        openGraph: {
+            title: ogTitle,
+            description: ogDescription,
+            url: ogUrl,
+            images: [
+                {
+                    url: ogImage,
+                    width: image_width,
+                    height: image_height,
+                    alt: ogTitle,
+                }
+            ],
+            type: 'website'
+        },
+        robots: (noindex === 1 || s !== "1") ? 'noindex, follow' : 'index, follow',
+        alternates: {
+            canonical: ogUrl,
+        },
+        icons: {
+            icon: [
+                { url: "/q-logo-light-42.png", media: "(prefers-color-scheme: light)" },
+                { url: "/q-logo-dark-42.png", media: "(prefers-color-scheme: dark)" }
+            ],
+            shortcut: [
+                { url: "/q-logo-light-512.png", media: "(prefers-color-scheme: light)" },
+                { url: "/q-logo-dark-512.png", media: "(prefers-color-scheme: dark)" }
+            ],
+        },
+    };
+}
